CODE RUB: Remove Storage Validations from Students Retrieve All
