+++ OtripleS.Web.Api.Tests.Unit/Services/Foundations/Students/StudentServiceTests.Exceptions.RetrieveAll.cs
@@ -1,0 +1,80 @@
+﻿// ---------------------------------------------------------------
+// Copyright (c) Coalition of the Good-Hearted Engineers
+// FREE TO USE AS LONG AS SOFTWARE FUNDS ARE DONATED TO THE POOR
+// ---------------------------------------------------------------
+
+using System;
+using Moq;
+using OtripleS.Web.Api.Models.Students.Exceptions;
+using Xunit;
+
+namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Students
+{
+    public partial class StudentServiceTests
+    {
+        [Fact]
+        public void ShouldThrowDependencyExceptionOnRetrieveAllIfSqlExceptionOccursAndLogIt()
+        {
+            // given
+            var sqlException = GetSqlException();
+
+            var expectedStudentDependencyException =
+                new StudentDependencyException(sqlException);
+
+            this.storageBrokerMock.Setup(broker =>
+                broker.SelectAllStudents())
+                    .Throws(sqlException);
+
+            // when . then
+            Assert.Throws<StudentDependencyException>(() =>
+                this.studentService.RetrieveAllStudents());
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectAllStudents(),
+                    Times.Once);
+
+            this.loggingBrokerMock.Verify(broker =>
+                broker.LogCritical(It.Is(SameExceptionAs(
+                    expectedStudentDependencyException))),
+                        Times.Once);
+
+            this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+        }
+
+        [Fact]
+        public void ShouldThrowServiceExceptionOnRetrieveAllIfExceptionOccursAndLogIt()
+        {
+            // given
+            var serviceException = new Exception();
+
+            var failedStudentServiceException =
+                new FailedStudentServiceException(serviceException);
+
+            var expectedStudentServiceException =
+                new StudentServiceException(failedStudentServiceException);
+
+            this.storageBrokerMock.Setup(broker =>
+                broker.SelectAllStudents())
+                    .Throws(serviceException);
+
+            // when . then
+            Assert.Throws<StudentServiceException>(() =>
+                this.studentService.RetrieveAllStudents());
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectAllStudents(),
+                    Times.Once);
+
+            this.loggingBrokerMock.Verify(broker =>
+                broker.LogError(It.Is(SameExceptionAs(
+                    expectedStudentServiceException))),
+                        Times.Once);
+
+            this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+        }
+    }
+}
