+++ OtripleS.Web.Api.Tests.Unit/Services/Foundations/Students/StudentServiceTests.Validations.RetrieveById.cs
@@ -4,10 +4,7 @@
 // ---------------------------------------------------------------
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
-using FluentAssertions;
 using Moq;
 using OtripleS.Web.Api.Models.Students;
 using OtripleS.Web.Api.Models.Students.Exceptions;
@@ -90,41 +87,6 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Students
 
             this.storageBrokerMock.Verify(broker =>
                 broker.SelectStudentByIdAsync(inputStudentId),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-            this.storageBrokerMock.VerifyNoOtherCalls();
-        }
-
-        [Fact]
-        public void ShouldLogWarningOnRetrieveAllWhenStudentsWasEmptyAndLogIt()
-        {
-            // given
-            IQueryable<Student> emptyStorageStudents = new List<Student>().AsQueryable();
-            IQueryable<Student> expectedStudents = emptyStorageStudents;
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.SelectAllStudents())
-                    .Returns(expectedStudents);
-
-            // when
-            IQueryable<Student> actualStudents =
-                this.studentService.RetrieveAllStudents();
-
-            // then
-            actualStudents.Should().BeEquivalentTo(emptyStorageStudents);
-
-            this.loggingBrokerMock.Verify(broker =>
-                broker.LogWarning("No students found in storage."),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.Verify(broker =>
-                broker.GetCurrentDateTime(),
-                    Times.Never);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectAllStudents(),
                     Times.Once);
 
             this.dateTimeBrokerMock.VerifyNoOtherCalls();
