+++ OtripleS.Web.Api/Services/Foundations/Students/StudentService.Validations.cs
@@ -4,7 +4,6 @@
 // ---------------------------------------------------------------
 
 using System;
-using System.Linq;
 using OtripleS.Web.Api.Models.Students;
 using OtripleS.Web.Api.Models.Students.Exceptions;
 
@@ -182,14 +181,6 @@ namespace OtripleS.Web.Api.Services.Foundations.Students
             if (student is null)
             {
                 throw new NullStudentException();
-            }
-        }
-
-        private void ValidateStorageStudents(IQueryable<Student> storageStudents)
-        {
-            if (!storageStudents.Any())
-            {
-                this.loggingBroker.LogWarning("No students found in storage.");
             }
         }
